CDC-20 Refactor: useSocialSignin.ts
